refactor(application,infrastructure): type mode/status params instead of raw int
Mechanical cleanup, lowest-risk phase of the explicit-typing pass:

- ILeaderboardStore/RedisLeaderboardStore: mode params are GameMode,
  cast to int only inside the Redis key-building helpers (key strings
  unchanged - same numeric value, just moved the cast site).
- PlayerSession.ModeStats/ScoreSubmissionUseCase/OsuLoginUseCase:
  Dictionary<int, CachedPlayerStats> -> Dictionary<GameMode, ...>,
  dropping the (int)/cast-back-and-forth at every call site.
- BeatmapInfoService: deleted a private RankedStatus->int dictionary
  that duplicated RankedStatusExtensions.OsuApi() field-for-field.
- DirectSearchService: named the "any mode"/"any ranked status" -1/4
  sentinels and the duplicated page-size-100 literal (shared with
  DirectSearchResponseFormatter via one internal const).
- MySqlChannelRepository: parameterized `auto_join = 1` instead of a
  literal in the SQL string.
- MySqlRelationshipRepository: added TypeFromColumn as the read-side
  counterpart to the existing TypeColumn, so the "friend"/"block"
  string encoding has one source of truth instead of two independent
  literal sites.

This is the last phase of the explicit-typing pass (see the plan doc
for the full 5-phase breakdown and what was deliberately left alone:
ScoreSubmission/LoginDataParser's positional field parsing, the two
different legacy boolean-string encodings, MySqlLogRepository's
single-use action string - all skipped as speculative abstractions
over working code, per CLAUDE.md's Simplicity First).

+++ src/OpenOsuTournament.Bancho.Application/UseCases/Scores/ScoreSubmissionUseCase.cs
@@ -123,7 +123,7 @@ public sealed class ScoreSubmissionUseCase(
                 // discarded for now; the score itself still counts. See note.md.
                 replayData = null;
 
-            var previousStats = player.ModeStats.GetValueOrDefault((int)score.Mode) ??
+            var previousStats = player.ModeStats.GetValueOrDefault(score.Mode) ??
                                 new CachedPlayerStats(0, 0, 0, 0, 0, 0, 0, 0);
             var updatedStats = ScoreStatsCalculator.ApplyScoreStats(score, previousStats);
             var shouldUpdateRank =
@@ -149,16 +149,16 @@ public sealed class ScoreSubmissionUseCase(
 
             if (shouldUpdateRank)
             {
-                await leaderboardStore.AddToGlobalLeaderboardAsync(player.Id, (int)score.Mode, updatedStats.Rscore,
+                await leaderboardStore.AddToGlobalLeaderboardAsync(player.Id, score.Mode, updatedStats.Rscore,
                     cancellationToken);
                 await leaderboardStore.AddToCountryLeaderboardAsync(
-                    player.Id, (int)score.Mode, player.Geoloc.CountryAcronym, updatedStats.Rscore, cancellationToken);
+                    player.Id, score.Mode, player.Geoloc.CountryAcronym, updatedStats.Rscore, cancellationToken);
                 var newRank =
-                    await leaderboardStore.FetchGlobalRankAsync(player.Id, (int)score.Mode, cancellationToken);
+                    await leaderboardStore.FetchGlobalRankAsync(player.Id, score.Mode, cancellationToken);
                 if (newRank is not null) updatedStats = updatedStats with { Rank = newRank.Value };
             }
 
-            player.ModeStats[(int)score.Mode] = updatedStats;
+            player.ModeStats[score.Mode] = updatedStats;
 
             if (replayData is not null) await replayStorage.WriteAsync(scoreId, replayData, cancellationToken);
